Make this mbed program.
I made this into a mbed program by doing 'mbed new .' in the empty repo. This took a while as the mbed-os repo was downloaded.

'mbed-os.lib' got added by the 'mbed new .' command, I didn't do it. I think it is just a link to the source of mbed-os.