Select compiler using local settings.
When I installed mbed CLI using the Windows installer the GCC ARM compiler also got installed.

But mbed doesn't find it automatically.  There are various ways to sort this:
   * Change mbed_settings.py (what I done here).
   * Change the path.
   * Use 'mbed config'.

This way is good because it captures the compiler version in the trpo but bad because everyone has to have the compiler in the same place.