Update local mbed config.
It's still not clear to me that I should have committed this file.

'.mbed' changed when 'mbed compile -m DISCO_L072CZ_LRWAN1 -t GCC_ARM' completed successfully. It appears to be remembering the options.

Whether or not we really want this remains to be seen, I guess it depends on how many targets the project supports.