Created some C++14 profiles.
I haven't found a way of adjust the compiler/linker options. I copied these files from 'mbed-os\tools\profiles' because changing them there felt like the wrong thing to do.

I like C++14, our compiler supports it so why not use it?

Perhaps because this is an embedded project and there's a negative impact on our executable size of something.

Arm decided to use C++ for mbed OS so I would have thought using the latest C++ was the correct thing to do.