Don't compile all the file in mbed-os.
mbed-os is massive and it takes ages to do a full compile even though most of it isn't used.

It turns out there is a way to not compile files that aren't being used '.mbedignore'.

I've created '.mbedignore' and added a few directories but not really enough to make any difference to the build time. I though about recursively listing all the folders and then adding them back in but didn't.