Added some example unit tests.
When I first saw the unit testing mentioned it tweaked my interest so I've done some investigation.

I was hoping for unit tests that could run on the host but that's not the case. This would probably be much more difficult and require mocking a whole load of stuff.

The unit tests run on the target but there is some infrastructure for running them. I don't have a target at the moment and it is possible that it doesn't run automatically, it probably depends on the debug probe.

More of a problem is the inability to deal with the real application 'main'. This problem is mentioned under known issues here https://docs.mbed.com/docs/mbed-os-handbook/en/5.2/advanced/testing/. I've tried to solve the problem with a macro and some conditionally compiled could but '-DUNIT_TEST' must be added to the test build command, i.e. 'mbed test --compile --tests-by-name tests-mbed-test-case-* -DUNIT_TEST'.