Simplifie '.mbedignore'.
When I first created '.mbedignore' I got the impressive from the documentation that it wasn't recursive, hence I started listing all the folders.

It turns out '/*' works and ignores the folders below so this makes things much simpler.

The 'mbed-test' program doesn't use the RTOS or anything from 'frameworks' but they are used in the test build hence they cannot be ignored.